+++ test/Ast/StatementSpec.hs
@@ -5,7 +5,7 @@ module Ast.StatementSpec ( spec ) where
 import Control.Applicative ( Alternative(empty) )
 import Test.Hspec
 
-import Ast.Expression ( Expression(Int32) )
+import Ast.Expression ( Expression(Int64) )
 import Ast.Identifier ( toIdentifier )
 import Ast.Statement ( Statement(..) )
 import Ast.Type ( Type(..) )
@@ -38,19 +38,19 @@ spec = do
 
             it "parses return statement" $ do
                 let mResult = tryParser (parse :: Parser Statement) "return 124;"
-                mResult `shouldBe` pure (Return (Int32 124), read "")
+                mResult `shouldBe` pure (Return (Int64 124), read "")
 
             it "parses return statement and leaves the rest" $ do
                 let mResult = tryParser (parse :: Parser Statement) "return 124;c"
-                mResult `shouldBe` pure (Return (Int32 124), read "c")
+                mResult `shouldBe` pure (Return (Int64 124), read "c")
 
             it "parses expression statement" $ do
                 let mResult = tryParser (parse :: Parser Statement) "124;"
-                mResult `shouldBe` pure (Expression (Int32 124), read "")
+                mResult `shouldBe` pure (Expression (Int64 124), read "")
 
             it "parses expression statement and leaves the rest" $ do
                 let mResult = tryParser (parse :: Parser Statement) "124;stuff"
-                mResult `shouldBe` pure (Expression (Int32 124), read "stuff")
+                mResult `shouldBe` pure (Expression (Int64 124), read "stuff")
 
             it "parses declaration statement without assignment" $ do
                 let mResult = tryParser (parse :: Parser Statement) "int stuff;"
@@ -62,19 +62,19 @@ spec = do
 
             it "parses declaration statement with assignment" $ do
                 let mResult = tryParser (parse :: Parser Statement) "int stuff = 0;"
-                mResult `shouldBe` pure (Declaration Int (toIdentifier "stuff") (Just (Int32 0)), read "")
+                mResult `shouldBe` pure (Declaration Int (toIdentifier "stuff") (Just (Int64 0)), read "")
 
             it "parses declaration statement wit assignment and leaves the rest" $ do
                 let mResult = tryParser (parse :: Parser Statement) "int stuff = 0;ccc"
-                mResult `shouldBe` pure (Declaration Int (toIdentifier "stuff") (Just (Int32 0)), read "ccc")
+                mResult `shouldBe` pure (Declaration Int (toIdentifier "stuff") (Just (Int64 0)), read "ccc")
 
         describe "PrettyPrint" $ do
             it "should render return statement" $ do
-                let statement = Return $ Int32 124
+                let statement = Return $ Int64 124
                 render statement `shouldBe` "RETURN 124"
 
             it "should render expression statement" $ do
-                let statement = Expression $ Int32 124
+                let statement = Expression $ Int64 124
                 render statement `shouldBe` "124"
 
             it "should render declaration statement without assignment" $ do
@@ -82,5 +82,5 @@ spec = do
                 render statement `shouldBe` "INT x"
 
             it "should render declaration statement with assignment" $ do
-                let statement = Declaration Int (toIdentifier "x") $ Just $ Int32 124
+                let statement = Declaration Int (toIdentifier "x") $ Just $ Int64 124
                 render statement `shouldBe` "INT x = 124"
